Get the graph form lines out of the graph form principal page
