+++ apps/site/app/containers/Compare/FormCompare/styles.module.less
@@ -7,39 +7,10 @@
     border-radius: 5px;
     color: white;
   
-    .variablesLabelContainer{
-      font-size: 18px;
-      font-weight: bold;
-      text-align: left;
-      margin-left: 15%;
-      margin-top: 50px;
-      margin-bottom: 30px;
-    }
-  
-    .variablesLabelValueContainer{
-      font-size: 16px;
-      font-weight: bold;
-      text-align: right;
-      padding-right: 10%;
-    }
-  
-    input[type=range]{
-        max-width: 350px;
-        min-width: 250px;
-    }
-  
-    .variablesScaleContainer{
-        font-size: 10px;
-        margin-top: -20px;
-        margin-bottom: 30px;
-    }
-  
     .buttonContainer{
       bottom: 10px;
       position:relative;
       margin-top: 80px;
     }
-  
-    .variablesValueContainer{
-    }
+ 
   }
